bluetooth/requestLEScan/radio-not-present.https.html

<!DOCTYPE html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-vendor.js"></script>
<script src="/bluetooth/resources/bluetooth-helpers.js"></script>
<script>
'use strict';
const test_desc = 'requestLEScan reject with NotFoundError if there is no BT radio present.';
const expected = new DOMException('Bluetooth adapter not available.',
    'NotFoundError');

if (!navigator.bluetooth.requestLEScan) {
  // requestLEScan is being tested behind a flag. When we get further along in the
  // implementation, we will remove this if-statement.
  done();
}

bluetooth_test(() => navigator.bluetooth.test.simulateCentral({state: 'absent'})
    .then(() => assert_promise_rejects_with_message(
        requestLEScanWithTrustedClick(),
        expected, 'Bluetooth adapter is not present.')),
    test_desc);
</script>
